requirements.txt
kafka-python==2.0.2pi
pandas==2.2.2
s3fs==2024.6.0